+++ racksdb/drawers/infrastructure.py
@@ -28,6 +28,22 @@ class InfrastructureDrawer(Drawer):
         self.rack_rows = []
         # List of racks used by the insfrastructure
         self.racks = []
+        # Calculated at draw time based on dimensions
+        self.ratio = 0
+
+    def _rack_row_width(self, rack_row):
+        """Return rack row width in mm"""
+        total = 0
+        for rack in rack_row.racks:
+            total += rack.type.width
+        return total
+
+    def _rack_row_height(self, rack_row):
+        """Return rack row height, ie. the maximum rack height in the row, in mm."""
+        row_max_height = 0
+        for rack in rack_row.racks:
+            row_max_height = max(row_max_height, rack.type.height)
+        return row_max_height
 
     def _rack_row_dl(self, row) -> ImagePoint:
         # sum height of all previous rows
@@ -36,12 +52,12 @@ class InfrastructureDrawer(Drawer):
             if _row is row:
                 break
             pos_y += (
-                int(_row.height * self.parameters.infrastructure.scale)
+                int(_row.height * self.ratio)
                 + self.parameters.row.label_offset
                 + self.parameters.rack.label_offset
                 + self.parameters.rack.offset
             )
-        pos_y += int(row.height * self.parameters.infrastructure.scale)
+        pos_y += int(row.height * self.ratio)
         return ImagePoint(self.parameters.margin.left, pos_y)
 
     def _rack_dl(self, row, rack) -> ImagePoint:
@@ -52,14 +68,14 @@ class InfrastructureDrawer(Drawer):
             for row_rack in rack.row.racks:
                 if row_rack.slot > rack.slot:
                     dl.x += (
-                        int(row_rack.type.width * self.parameters.infrastructure.scale)
+                        int(row_rack.type.width * self.ratio)
                         + self.parameters.rack.spacing
                     )
         else:
             for row_rack in rack.row.racks:
                 if row_rack.slot < rack.slot:
                     dl.x += (
-                        int(row_rack.type.width * self.parameters.infrastructure.scale)
+                        int(row_rack.type.width * self.ratio)
                         + self.parameters.rack.spacing
                     )
         dl.y += self.parameters.rack.label_offset + self.parameters.rack.offset
@@ -92,14 +108,14 @@ class InfrastructureDrawer(Drawer):
             + equipment_width_slot
             * equipment.type.width
             * (
-                int(equipment.rack.type.width * self.parameters.infrastructure.scale)
+                int(equipment.rack.type.width * self.ratio)
                 - 2 * self.parameters.rack.pane_width
             )
         )
         tl.y -= (
             (equipment.type.height + equipment_height_slot)
             * self.parameters.rack.u_height
-            * self.parameters.infrastructure.scale
+            * self.ratio
         )
 
         return tl
@@ -115,13 +131,11 @@ class InfrastructureDrawer(Drawer):
         tl = self._equipment_tl(row, rack, equipment)
 
         equipment_width = equipment.type.width * (
-            int(equipment.rack.type.width * self.parameters.infrastructure.scale)
+            int(equipment.rack.type.width * self.ratio)
             - 2 * self.parameters.rack.pane_width
         )
         equipment_height = int(
-            equipment.type.height
-            * self.parameters.rack.u_height
-            * self.parameters.infrastructure.scale
+            equipment.type.height * self.parameters.rack.u_height * self.ratio
         )
 
         colorset = self._find_equipment_colorset(equipment)
@@ -165,8 +179,8 @@ class InfrastructureDrawer(Drawer):
         # top left of rack
         dl = self._rack_dl(row, rack)
 
-        rack_width = rack.type.width * self.parameters.infrastructure.scale
-        rack_height = rack.type.height * self.parameters.infrastructure.scale
+        rack_width = rack.type.width * self.ratio
+        rack_height = rack.type.height * self.ratio
 
         # write rack name
         self.ctx.move_to(dl.x, dl.y - rack_height - self.parameters.rack.offset)
@@ -225,9 +239,7 @@ class InfrastructureDrawer(Drawer):
             "Sans", cairo.FONT_SLANT_NORMAL, cairo.FONT_WEIGHT_BOLD
         )
         self.ctx.set_font_size(14)
-        self.ctx.move_to(
-            dl.x, dl.y - int(row.height * self.parameters.infrastructure.scale)
-        )
+        self.ctx.move_to(dl.x, dl.y - int(row.height * self.ratio))
         self.ctx.show_text(f"row {row.name}")
 
         # iterate over the racks to draw racks in row
@@ -241,6 +253,12 @@ class InfrastructureDrawer(Drawer):
 
     def draw(self):
 
+        logger.debug(
+            "Maximum dimensions: %d %d",
+            self.parameters.dimensions.width,
+            self.parameters.dimensions.height,
+        )
+
         # Get list of racks and rows used by the infrastructure
         for part in self.infrastructure.layout:
             if part.rack not in self.racks:
@@ -248,34 +266,55 @@ class InfrastructureDrawer(Drawer):
             if part.rack.row not in self.rack_rows:
                 self.rack_rows.append(part.rack.row)
 
-        # Sum all rows maximum rack height to calculate image height
+        # Search maximum racks row width and sum all rows maximum rack height in mm
+        max_row_width = 0
+        max_row_racks = 0
         total_row_max_heights = 0
         for rack_row in self.rack_rows:
-            row_max_height = 0
-            for rack in rack_row.racks:
-                row_max_height = max(row_max_height, rack.type.height)
-            total_row_max_heights += row_max_height
-            # add attributes to row
-            rack_row.height = row_max_height
+            max_row_width = max(max_row_width, self._rack_row_width(rack_row))
+            max_row_racks = max(max_row_racks, len(rack_row.racks))
+            # add height attribute to row
+            rack_row.height = self._rack_row_height(rack_row)
+            total_row_max_heights += rack_row.height
 
-        # Find the maximum rack x to calculate image width
-        total_racks_widths = 0
-        for rack in self.racks:
-            dl = self._rack_dl(rack.row, rack)
-            x_tr = dl.x + int(rack.type.width * self.parameters.infrastructure.scale)
-            total_racks_widths = max(total_racks_widths, x_tr)
-
-        surface_width = total_racks_widths + self.parameters.margin.left
-        surface_height = (
-            2 * self.parameters.margin.top
-            + len(self.rack_rows)
-            * (
-                self.parameters.row.label_offset
-                + self.parameters.rack.label_offset
-                + self.parameters.rack.offset
-            )
-            + int(total_row_max_heights * self.parameters.infrastructure.scale)
+        logger.debug(
+            "total row heights: %d max row width: %d",
+            total_row_max_heights,
+            max_row_width,
         )
+
+        # Compute total white spaces in width and height
+        width_whitespace = (
+            self.parameters.margin.left * 2
+            + self.parameters.rack.spacing * (max_row_racks - 1)
+        )
+        height_whitespace = 2 * self.parameters.margin.top + len(self.rack_rows) * (
+            self.parameters.row.label_offset
+            + self.parameters.rack.label_offset
+            + self.parameters.rack.offset
+        )
+
+        # Deduce available space to draw in width and height
+        available_width = self.parameters.dimensions.width - width_whitespace
+        available_height = self.parameters.dimensions.height - height_whitespace
+        logger.debug(
+            "available draw area in pixels heights: %d width: %d",
+            available_height,
+            available_width,
+        )
+
+        # Compute final ratio
+        self.ratio = min(
+            available_width / max_row_width, available_height / total_row_max_heights
+        )
+        logger.debug("Final ratio is: %f", self.ratio)
+
+        # Compute final surface width and height
+        surface_width = int(self.ratio * max_row_width + width_whitespace)
+        surface_height = int(self.ratio * total_row_max_heights + height_whitespace)
+
+        logger.debug("Final width: %d", surface_width)
+        logger.debug("Final height: %d", surface_height)
 
         self.init_ctx(surface_width, surface_height)
         self._draw_infrastructure()